+++ doc/examples/gerchberg_saxton.py
@@ -7,9 +7,9 @@ Created on Thu Aug  4 15:44:04 2022
 
 import numpy as np
 
-from lightprop.algorithms.gs import GerchbergSaxton
 from lightprop.calculations import gaussian, get_gaussian_distribution
 from lightprop.lightfield import LightField
+from lightprop.optimization.gs import GerchbergSaxton
 from lightprop.propagation.params import PropagationParams
 from lightprop.visualisation import Plotter, PlotTypes
